fix test

+++ tests/nodes/maxpool_2d_constraint_index/output_0.cairo
@@ -1,9 +1,9 @@
 use core::array::{ArrayTrait, SpanTrait};
 use orion::operators::tensor::{TensorTrait, Tensor};
-use orion::operators::tensor::{I32Tensor, I32TensorAdd};
+use orion::operators::tensor::{U32Tensor, U32TensorAdd};
 use orion::numbers::NumberTrait;
 
-fn output_0() -> Tensor<i32> {
+fn output_0() -> Tensor<usize> {
     let mut shape = ArrayTrait::<usize>::new();
     shape.append(1);
     shape.append(1);
